requirements.txt
pandas==2.0.3
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.2.0
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
PyYAML==6.0
python-dotenv==1.0.0
seaborn==0.13.2
matplotlib==3.8.0
xgboost==2.0.3